hclwrite: Make File and Tokens implement io.WriterTo
There isn't any strong reason for this -- they don't implement io.Reader
and so can't be used in places where a Reader+WriterTo is expected, like
io.Copy -- but go lint thinks that anything called WriteTo with an
io.Writer argument is an attempt to implement WriterTo and so this just
shuts up the linter.

+++ hclwrite/ast.go
@@ -22,7 +22,7 @@ func (f *File) Body() *Body {
 //
 // The tokens first have a simple formatting pass applied that adjusts only
 // the spaces between them.
-func (f *File) WriteTo(wr io.Writer) (int, error) {
+func (f *File) WriteTo(wr io.Writer) (int64, error) {
 	tokens := f.inTree.children.BuildTokens(nil)
 	format(tokens)
 	return tokens.WriteTo(wr)
